Moved CreateTestTaker to Test Taker API
